refactor(react): extract Component activation into component.ts
Move the Component-prototype wiring out of the grab-bag runtime.ts into a
dedicated component.ts (pairing with the existing component.test.tsx): the
deprecated host-JSX member declare-module, the state/context proto descriptors,
bootstrap, the per-instance render host, subcomponents, and the Component.on
prepare handler.

runtime.ts is now just what its name says - the Runtime seam, hook primitives
(useFactory/useReady/useHook), the standalone use hook, and the barrel - plus
side-effect imports of ./component and ./state.get/.state.use to trigger
activation. State.get/use stay self-assigning in their own modules; wiring them
elsewhere would be cross-concern.
